+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Update server integration tests for new validation and convention-files
 ---
 
 ## [0.20.0] - 2026-03-22